Organize test suite
