+++ backend/tests/integration/post-route-test.js
@@ -1,6 +1,6 @@
 const express = require("express");
-const Post = require('../schema/post-schema');
-const User = require('../schema/user-schema');
+const Post = require('../../schema/post-schema');
+const User = require('../../schema/user-schema');
 const mongoose = require('mongoose');
 const expect = require('chai').expect;
 const axios = require('axios');
@@ -8,9 +8,9 @@ require("dotenv").config();
 
 const cors = require('cors')
 const bodyParser = require('body-parser');
-const apiRouter = require("../route/api-route")
+const apiRouter = require("../../route/api-route")
 const passport = require("passport")
-require("../util/passport")(passport)
+require("../../util/passport")(passport)
 
 const username = 'completelyNewUsername';
 const email = 'goodBoi@email.com';
